refactor: tool-bar reuse
